Improve chart performance

+++ PV6900.Wpf/Models/TimeSpanAmperePoint.cs
@@ -9,8 +9,8 @@ namespace PV6900.Wpf.Models
 {
     public class TimeSpanAmperePoint:BindableBase
     {
-        public long TimeSpan { get => _timeSpan; set => SetProperty(ref _timeSpan, value); }
-        private long _timeSpan;
+        public double TimeSpan { get => _timeSpan; set => SetProperty(ref _timeSpan, value); }
+        private double _timeSpan;
         public double Ampere { get => _ampere; set => SetProperty(ref _ampere, value); }
         private double _ampere;
     }
