+++ PV6900.Wpf/Models/TimeSpanVoltaPoint.cs
@@ -9,8 +9,8 @@ namespace PV6900.Wpf.Models
 {
     public class TimeSpanVoltaPoint : BindableBase
     {
-        public long TimeSpan { get => _timeSpan; set => SetProperty(ref _timeSpan, value); }
-        private long _timeSpan;
+        public double TimeSpan { get => _timeSpan; set => SetProperty(ref _timeSpan, value); }
+        private double _timeSpan;
         public double Volta { get => _volta; set => SetProperty(ref _volta, value); }
         private double _volta;
     }
